JUnit Test cases for Inventory API with Java Code Coverage 23% V1

+++ src/main/java/com/iemr/inventory/repo/stockadjustment/StockAdjustmentItemRepo.java
@@ -35,7 +35,7 @@ import com.iemr.inventory.data.stockadjustment.StockAdjustmentItem;
 import com.iemr.inventory.data.stockadjustment.StockAdjustmentItemDraft;
 
 @Repository
-public interface StockAdjustmentItemRepo extends CrudRepository<StockAdjustmentItem, Integer> {
+public interface StockAdjustmentItemRepo extends CrudRepository<StockAdjustmentItem, Long> {
 
 	@Transactional
 	@Modifying
